backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
drf-spectacular==0.26.5
dj-database-url==2.0.0
djangorestframework-simplejwt==5.2.2